refactor: moved GradientLink to UI
