+++ routes/projects.tsx
@@ -4,7 +4,7 @@ import { CustomHead } from "../components/base/CustomHead.tsx";
 import { Base } from "../components/base/Base.tsx";
 //? Default styled header
 import { StyledHeader } from "../components/UI/StyledHeader.tsx";
-import { GradientLink } from "../components/base/GradientLink.tsx";
+import { GradientLink } from "../components/UI/GradientLink.tsx";
 //? Navigation Buttons to go back to the previous page or to the next article
 import { BlogNavigationButtons } from "../components/blog/BlogNavigationButtons.tsx";
 //? Component to create a column grid of projects created
