+++ routes/work.tsx
@@ -7,7 +7,7 @@ import { StyledHeader } from "../components/UI/StyledHeader.tsx";
 //? Navigation Buttons to go back to the previous page or to the next article
 import { BlogNavigationButtons } from "../components/blog/BlogNavigationButtons.tsx";
 //? A HTML Link component to pre-format links and reduce boiletplate
-import { GradientLink } from "../components/base/GradientLink.tsx";
+import { GradientLink } from "../components/UI/GradientLink.tsx";
 
 export default function Home() {
   return (
